requirements.txt
# ═══════════════════════════════════════════════════════════════════
# Obscuras Campaign Manager - Dependencies
# Python 3.10+ required
# ═══════════════════════════════════════════════════════════════════

# Core
PyYAML>=6.0

# GUI Framework
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0  # Für HTML-Vorschau

# Database
SQLAlchemy>=2.0.0
alembic>=1.13.0  # Datenbank-Migrationen

# Data Processing
pandas>=2.1.0  # CSV/Excel Import
openpyxl>=3.1.0  # Excel-Support

# E-Mail
email-validator>=2.1.0  # E-Mail-Validierung
dnspython>=2.4.0  # MX-Record Prüfung

# Security
cryptography>=41.0.0  # Passwort-Verschlüsselung
keyring>=24.3.0  # Sichere Credential-Speicherung

# Utilities
python-dateutil>=2.8.0
Jinja2>=3.1.0  # Template-Engine
rich>=13.7.0  # Schöne CLI-Ausgabe

# Development (optional)
# pytest>=7.4.0
# black>=23.0.0
